refactor: 파일 정리, 모달 위치 수정
